requirements.txt
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
bokeh==2.0.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.0
Keras-Preprocessing==1.1.2
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.17
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
threadpoolctl==2.0.0
tornado==6.0.4
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
